requirements.txt
# Core Dependencies
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.6.0
pydantic-settings>=2.1.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.7
python-dotenv>=1.0.0

# Markdown Processing
markdown2>=2.4.12
pypandoc>=1.13.0

# Document Conversion
pdfkit>=1.0.0
python-docx>=1.1.0
html2image>=2.0.3
weasyprint>=60.2
reportlab>=4.0.8

# HTTP and API
requests>=2.31.0
httpx>=0.26.0
aiofiles>=23.2.1

# Database and Migration
alembic>=1.13.1
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Testing and Development
pytest>=7.4.4
pytest-asyncio>=0.23.3
pytest-cov>=4.1.0
black>=24.1.0
isort>=5.13.2
flake8>=7.0.0
mypy>=1.8.0

# Security
cryptography>=42.0.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.2

# Monitoring and Logging
structlog>=24.1.0
python-json-logger>=2.0.7

# Image Processing
Pillow>=10.2.0
cairosvg>=2.7.1

# System Dependencies (commented as they are installed via package manager)
# pandoc >= 3.1.3
# wkhtmltopdf >= 0.12.6
# texlive-xetex
# texlive-fonts-recommended
# libpangocairo-1.0-0
# fonts-liberation
# xvfb
# chromium-browser

# Conversion tools
pypandoc==1.12
pdfkit==1.0.0
python-docx==0.8.11
html2image==2.0.1
openpyxl==3.1.2
csvkit==1.1.1 